+++ src/main/java/com/cs/orderbookmanagement/repository/OrderBookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderBookRepository extends JpaRepository<OrderBook, Integer> {
+public interface OrderBookRepository extends JpaRepository<OrderBook, Long> {
 
 
 }
